Work protocol selection into command line arguments
Leave the Connection to mean the transport connection, but add
a Protocol parameter to Peer and Listen structures to signify the
protocol.

Don't rely on FromStr impls to select connection address:
as implemented, the parsing rules were ambiguous. It's better to
make it explicit in option names and/or syntax.