Log a warning when block solicitation fails
Quoting Nicolas on PR#357:
This is not really an error, i agree. But not being able to solicit a
block might be problematic, and this is a warning to something that may
become a problem down the line (not being able to sync the blockchain
with the rest of the world).