network: Pull and push missing chain headers
Update chain-deps to add support for PushHeaders, PullHeaders,
and the solicitation for headers of blocks missing in the blockchain.

Handle the case when the blockchain triage finds that a block announced
by a network peer is missing parents in local storage. The solicitation
is sent to the network task, which handles it asymmetrically:
on the client side, the PullHeaders request is issued, and on the
server side, a Missing solicitation event is sent down the subscription
stream. The Missing event handling is implemented on the client side.

Not implemented is processing of the response stream of PullHeaders
or the request stream of PushHeaders. A new type of intercom message(s)
will be needed for this.

Also change the reply handle of intercom::ClientMsg::GetHeadersRange
to stream the headers.