CircleCI: Streamline the workflow
In the build jobs, only build the binaries that are needed for
the integration tests. The dependency build artefacts are still
persisted and reused for the test jobs.

In the test jobs, build the tests in a separate step to reduce
non-test output in the test step.

Clean up restore_cache steps: only precise cache hits are used
for test jobs, as verified by the --frozen option given to cargo.