blockchain: Handle out-of-order network blocks
If the block does not have a locally stored parent, put it
into the CandidateRepo's cache, but only if its header is already
attested there from a previously processed header chain.

This is to be complemented by picking up and applying cached blocks
whenever the candidate branch is to be advanced.