Cancellable future to connect client
A future that drives client connection and stream subscriptions,
that needs to be cancellable at any point in the process.

Reworked client::connect to return the abovementioned future
and the cancellation/result handle yielding PeerComms when
the client connects.

PeerMap keeps track of the ConnectHandle for a peer while the client
is connecting, providing accessors to peer's communication handles
possibly updated after the connection is established, as well as
methods called when a server subscription request comes in, which
cancel the pending client connection and return the comm handles
cleaned up from pending items in expectation of server-side
subscriptions.

PeerComms::update merges communication handles from a
new peer into an existing entry, taking into account any
pending items.