Don't thrash slow connections in propagation
When a peer is slow to connect to and there is another item
to propagate, buffer the last propagation item in the PeerComms
state rather than failing the connection and attempting to connect
again. Otherwise, a connection could be thrashed in and out of
the connecting state for as long as there are new items for which
it is selected to propagate to.