PoC preferred list for poldercast layer
this PR contains a preferred list of nodes that can be used to build
a special layer of extra nodes that we want to keep in constant connections
with.

This is going to be useful for IOHK or operators with multiple nodes to
make sure their nodes are constantly communicating with each others.