Reworked stream replies to prevent ghost streams
Our handling of outbound streams has been lacking in error
reporting: when handling of a stream request or solicitation
was rejected due to, say, wrong parameters, the error was sent
in the stream. With upload streams, it now means the error cannot
be propagated to the server in the client streaming request,
resulting in a request with an empty outbound stream.
With server streaming responses, we haven't been doing the right thing
either: instead of sending a proper error response, we sent an OK in the
headers, then an error in the trailers.

Rework intercom::stream_reply to return a future, which
provides the missing error reporting stage.

Opportunistically, clean up the pump code in blockchain::storage.