Keep track of demoted nodes in quarantine
Poldercast alone is not sufficient for use to implement
a quarantine like it existed before, since there would
be no way of promoting a peer back to the good nodes.
Add a struct to keep track of quarantined nodes and
re-add them for consideration in the topology after
quarantine time elapsed.