Move topology to its own task
Topology was handled directly by the network task, and that
, apart from not being very clean, was causing prolems with
the new update since topolgy cannot be shared easiliy anymore.

The 'p2p/peers' and 'topology' layers have to talk to each other
to exchange information (mostly topology -> peers)). At the moment
each has its own representation for a node and the only thing that
is taken into consideration from the other layer is the address of
the node. In the future we would likely want to add verification
of identities that would require exchanging (and defining a way to
do so) node public keys. Since the precise way we will use to do
so has not been decided so, this commit does not try to merge the
two representations together.