Propagate to one peer at a time
Since we need to have this workaround until we decide how peers
and topology layers should exchange information about external
nodes, change the signature of propagation function to propagate
to a single address to avoid having to create Vec unnecessarily.
By doing this we lose the ability of tracking if a node can be
reached from multiple addresses, but it's not a top priority and
should be quite uncommon anyway (and even if that happens is not
a problem per se).