ci: remove draft release on non-build failures
With the previous condition releases were removed only when one of
`build_assets` jobs was failing. However, this condition would not apply
if the failure occurred before launching this job. Instead, we check the
status of `publish_release` to be `success`. Except for a failure during
publishing, if one of the earlier jobs failed, the status of
`publish_release` would be `cancelled` which also trigger
`remove_failed_release`.