Handle tip updates in ad-hoc component
Refactor tip updates in a separate component to simplify code
and fix current inconsistencies.

Concurrent tip updates have resulted in torn-updates and incorrect
computations. To avoid locking the tip for the duration of the tip
selection algorithm, let's synchronize by message passing instead.
As a result, the tip cannot be updated anymore without a TipUpdater
handle.

In addition, this allow us to separate further code by function and
make it more readable and avoid passing useless structs around in
the program.

Lastly, create a branch even for refs that do not win branch
selection.